+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-# Changelog
+﻿# Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 2.2.0 - 2020-04-24
 
 * Added: New cmdlets Get-IBHFileEncoding and Set-ModuleVersion
 * Added: Check file encoding during the meta tests